+++ python/python/tests/test_transaction.py
@@ -2,7 +2,7 @@
 
 import time
 from datetime import date, datetime, timedelta, timezone
-from typing import List, Union, cast
+from typing import List, Optional, Union, cast
 
 import pytest
 from glide import RequestError
@@ -747,14 +747,28 @@ async def transaction_test(
 
 @pytest.mark.asyncio
 class TestTransaction:
+
+    async def exec_transaction(
+        self, glide_client: TGlideClient, transaction: BaseTransaction
+    ) -> Optional[List[TResult]]:
+        """
+        Exec a transaction on a client with proper client. Casts are required to satisfy `mypy`.
+        """
+        if isinstance(glide_client, GlideClient):
+            return await cast(GlideClient, glide_client).exec(
+                cast(Transaction, transaction)
+            )
+        else:
+            return await cast(GlideClusterClient, glide_client).exec(
+                cast(ClusterTransaction, transaction)
+            )
+
     @pytest.mark.parametrize("cluster_mode", [True])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
-    async def test_transaction_with_different_slots(self, redis_client: TGlideClient):
-        transaction = (
-            Transaction()
-            if isinstance(redis_client, GlideClient)
-            else ClusterTransaction()
-        )
+    async def test_transaction_with_different_slots(
+        self, redis_client: GlideClusterClient
+    ):
+        transaction = ClusterTransaction()
         transaction.set("key1", "value1")
         transaction.set("key2", "value2")
         with pytest.raises(RequestError, match="CrossSlot"):
@@ -771,7 +785,7 @@ class TestTransaction:
         )
         transaction.custom_command(["HSET", key, "foo", "bar"])
         transaction.custom_command(["HGET", key, "foo"])
-        result = await redis_client.exec(transaction)
+        result = await self.exec_transaction(redis_client, transaction)
         assert result == [1, b"bar"]
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
@@ -787,7 +801,7 @@ class TestTransaction:
         )
         transaction.custom_command(["WATCH", key])
         with pytest.raises(RequestError) as e:
-            await redis_client.exec(transaction)
+            await self.exec_transaction(redis_client, transaction)
         assert "WATCH inside MULTI is not allowed" in str(
             e
         )  # TODO : add an assert on EXEC ABORT
@@ -806,7 +820,7 @@ class TestTransaction:
         transaction.custom_command(["INCR", key])
         transaction.custom_command(["DISCARD"])
         with pytest.raises(RequestError) as e:
-            await redis_client.exec(transaction)
+            await self.exec_transaction(redis_client, transaction)
         assert "EXEC without MULTI" in str(e)  # TODO : add an assert on EXEC ABORT
         value = await redis_client.get(key)
         assert value == b"1"
@@ -818,7 +832,7 @@ class TestTransaction:
         transaction = BaseTransaction()
         transaction.custom_command(["INCR", key, key, key])
         with pytest.raises(RequestError) as e:
-            await redis_client.exec(transaction)
+            await self.exec_transaction(redis_client, transaction)
         assert "wrong number of arguments" in str(
             e
         )  # TODO : add an assert on EXEC ABORT
@@ -864,7 +878,7 @@ class TestTransaction:
         result2 = await client2.set(keyslot, "foo")
         assert result2 == OK
 
-        result3 = await redis_client.exec(transaction)
+        result3 = await self.exec_transaction(redis_client, transaction)
         assert result3 is None
 
         await client2.close()
@@ -969,7 +983,8 @@ class TestTransaction:
         transaction = ClusterTransaction() if cluster_mode else Transaction()
         transaction.set(key, "value").get(key).delete([key])
 
-        assert await redis_client.exec(transaction) == [OK, b"value", 1]
+        result = await self.exec_transaction(redis_client, transaction)
+        assert result == [OK, b"value", 1]
 
     # The object commands are tested here instead of transaction_test because they have special requirements:
     # - OBJECT FREQ and OBJECT IDLETIME require specific maxmemory policies to be set on the config
@@ -999,7 +1014,7 @@ class TestTransaction:
             transaction.config_set({maxmemory_policy_key: "allkeys-random"})
             transaction.object_idletime(string_key)
 
-            response = await redis_client.exec(transaction)
+            response = await self.exec_transaction(redis_client, transaction)
             assert response is not None
             assert response[0] == OK  # transaction.set(string_key, "foo")
             assert response[1] == b"embstr"  # transaction.object_encoding(string_key)
@@ -1024,7 +1039,7 @@ class TestTransaction:
         yesterday_unix_time = time.mktime(yesterday.timetuple())
         transaction = ClusterTransaction() if cluster_mode else Transaction()
         transaction.lastsave()
-        response = await redis_client.exec(transaction)
+        response = await self.exec_transaction(redis_client, transaction)
         assert isinstance(response, list)
         lastsave_time = response[0]
         assert isinstance(lastsave_time, int)
@@ -1033,7 +1048,7 @@ class TestTransaction:
     @pytest.mark.parametrize("cluster_mode", [True])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_lolwut_transaction(self, redis_client: GlideClusterClient):
-        transaction = Transaction()
+        transaction = ClusterTransaction()
         transaction.lolwut().lolwut(5).lolwut(parameters=[1, 2]).lolwut(6, [42])
         results = await redis_client.exec(transaction)
         assert results is not None